@@ -1,19 +1,20 @@
 """Utility functions for checking types and values. Inspired from MNE."""
 
+import logging
 import operator
 import os
 from pathlib import Path
+from typing import Any, Optional, Union
 
 import numpy as np
 
 
-def _ensure_int(item, item_name=None):
-    """
-    Ensure a variable is an integer.
+def _ensure_int(item: Any, item_name: Optional[str] = None) -> int:
+    """Ensure a variable is an integer.
 
     Parameters
     ----------
-    item : object
+    item : Any
         Item to check.
     item_name : str | None
         Name of the item to show inside the error message.
@@ -42,7 +43,7 @@ def _ensure_int(item, item_name=None):
 
 class _IntLike:
     @classmethod
-    def __instancecheck__(cls, other):
+    def __instancecheck__(cls, other: Any) -> bool:
         try:
             _ensure_int(other)
         except TypeError:
@@ -53,7 +54,7 @@ class _IntLike:
 
 class _Callable:
     @classmethod
-    def __instancecheck__(cls, other):
+    def __instancecheck__(cls, other: Any) -> bool:
         return callable(other)
 
 
@@ -65,9 +66,10 @@ _types = {
 }
 
 
-def _check_type(item, types, item_name=None):
-    """
-    Check that item is an instance of types.
+def _check_type(
+    item: Any, types: tuple, item_name: Optional[str] = None
+) -> Any:
+    """Check that item is an instance of types.
 
     Parameters
     ----------
@@ -122,9 +124,13 @@ def _check_type(item, types, item_name=None):
     return item
 
 
-def _check_value(item, allowed_values, item_name=None, extra=None):
-    """
-    Check the value of a parameter against a list of valid options.
+def _check_value(
+    item: Any,
+    allowed_values: tuple,
+    item_name: Optional[str] = None,
+    extra: Optional[str] = None,
+) -> Any:
+    """Check the value of a parameter against a list of valid options.
 
     Parameters
     ----------
@@ -169,3 +175,47 @@ def _check_value(item, allowed_values, item_name=None, extra=None):
         )
 
     return item
+
+
+def _check_verbose(verbose: Union[bool, str, int, None]) -> int:
+    """Check that the value of verbose is valid.
+
+    Parameters
+    ----------
+    verbose : bool | str | int | None
+        The verbosity level.
+
+    Returns
+    -------
+    verbose : int
+        The verbosity level as an integer.
+    """
+    logging_types = dict(
+        DEBUG=logging.DEBUG,
+        INFO=logging.INFO,
+        WARNING=logging.WARNING,
+        ERROR=logging.ERROR,
+        CRITICAL=logging.CRITICAL,
+    )
+
+    _check_type(verbose, (bool, str, "int", None), item_name="verbose")
+
+    if verbose is None:
+        verbose = logging.WARNING
+    elif isinstance(verbose, str):
+        verbose = verbose.upper()
+        _check_value(verbose, logging_types, item_name="verbose")
+        verbose = logging_types[verbose]
+    elif isinstance(verbose, bool):
+        if verbose:
+            verbose = logging.INFO
+        else:
+            verbose = logging.WARNING
+    elif isinstance(verbose, int):
+        if verbose <= 0:
+            raise ValueError(
+                "Argument 'verbose' can not be a negative integer, "
+                f"{verbose} is invalid."
+            )
+
+    return verbose
